Move package resolve tests
